Add Yast2::SystemService class
