+++ library/systemd/src/modules/systemd_service.rb
@@ -212,101 +212,6 @@ module Yast
         !socket.nil?
       end
 
-      alias_method :enable_service, :enable
-      private :enable_service
-
-      # Enable a service
-      #
-      # @param mode [Symbol] Start mode (:boot or :demand).
-      # @return [Boolean] true if the service was successfully enabled; false otherwise.
-      def enable
-        self.start_mode = :boot
-      end
-
-      alias_method :disable_service, :disable
-      private :disable_service
-
-      # Disable a service
-      #
-      # If the service has an associated socket, it is disabled too.
-      #
-      # @return [Boolean] true if the service was successfully disabled; false otherwise.
-      def disable
-        self.start_mode = :manual
-      end
-
-      alias_method :enabled_on_boot?, :enabled?
-      private :enabled_on_boot?
-
-      # Determine whether the service is enabled or not
-      #
-      # The service can be enable to be started on boot or on demand.
-      #
-      # @return [Boolean] true if the service is enabled; false otherwise.
-      def enabled?
-        start_mode != :manual
-      end
-
-      # Determine whether the service has an associated socket
-      #
-      # @return [Boolean] true if an associated socket exists; false otherwise.
-      def socket?
-        !socket.nil?
-      end
-
-      # Return the start mode
-      #
-      # See {#start_modes} to find out the supported modes for a given service (usually :on_boot,
-      # :manual and, in some cases, :on_demand).
-      #
-      # When the service (:on_boot) and the socket (:on_demand) are enabled, the start mode is translated
-      # to :on_boot.
-      #
-      # @return [Symbol] Start mode (:on_boot, :on_demand, :manual)
-      def start_mode
-        return :on_boot if enabled_on_boot?
-        return :on_demand if socket && socket.enabled?
-        :manual
-      end
-
-      # Set the service start mode
-      #
-      # See {#start_modes} to find out the supported modes for a given service (usually :on_boot,
-      # :manual and, in some cases, :on_demand).
-      #
-      # @see #start_modes
-      def start_mode=(mode)
-        if !start_modes.include?(mode)
-          raise ArgumentError, "Invalid start mode: '#{mode}' for service '#{name}'"
-        end
-
-        case mode
-        when :on_boot
-          enable
-          socket.disable
-        when :on_demand
-          disable
-          socket.enable
-        when :manual
-          disable
-          socket.disable
-        end
-      end
-
-      # Return the list of supported start modes
-      #
-      # * :on_boot:   The service will be started when the system boots.
-      # * :manual: The service is disabled and it will be started manually.
-      # * :on_demand: The service will be started on demand (using a Systemd socket).
-      #
-      # @return [Array<Symbol>] List of supported modes.
-      def start_modes
-        return @start_modes if @start_modes
-        @start_modes = [:on_boot, :manual]
-        @start_modes.insert(1, :on_demand) if socket?
-        @start_modes
-      end
-
     private
 
       def installation_system?
