+++ library/systemd/test/systemd_service_test.rb
@@ -144,187 +144,6 @@ module Yast
       end
     end
 
-    describe "#socket" do
-      it "returns nil if service does not have socket" do
-        service = SystemdService.find("sshd")
-        expect(service.socket).to eq nil
-      end
-
-      it "returns a socket that can start service" do
-        stub_services(service: "cups")
-        service = SystemdService.find("cups")
-        expect(service.socket).to be_a Yast::SystemdSocketClass::Socket
-      end
-    end
-
-    describe "#enabled?" do
-      subject(:service) { SystemdService.find("cups") }
-
-      before do
-        allow(service).to receive(:start_mode).and_return(start_mode)
-        stub_services(service: "cups")
-      end
-
-      context "when the start mode is :on_boot" do
-        let(:start_mode) { :on_boot }
-
-        it "returns true" do
-          expect(service).to be_enabled
-        end
-      end
-
-      context "when the start mode is :on_demand" do
-        let(:start_mode) { :on_demand }
-
-        it "returns true" do
-          expect(service).to be_enabled
-        end
-      end
-
-      context "when the start mode is :manual" do
-        let(:start_mode) { :manual }
-
-        it "returns false" do
-          expect(service).to_not be_enabled
-        end
-      end
-    end
-
-    describe "#start_mode" do
-      let(:enabled_on_boot?) { true }
-      let(:socket) { double("socket", enabled?: true) }
-
-      subject(:service) { SystemdService.find("cups") }
-
-      before do
-        stub_services(service: "cups")
-        allow(service).to receive(:socket).and_return(socket)
-        allow(service).to receive(:enabled_on_boot?).and_return(enabled_on_boot?)
-      end
-
-      context "when the service is enabled" do
-        it "returns :on_boot" do
-          expect(service.start_mode).to eq(:on_boot)
-        end
-      end
-
-      context "when the service is disabled" do
-        let(:enabled_on_boot?) { false }
-
-        context "but the associated socket is enabled" do
-          it "returns :on_demand" do
-            expect(service.start_mode).to eq(:on_demand)
-          end
-        end
-
-        context "and the socket is disabled" do
-          let(:socket) { double("socket", enabled?: false) }
-
-          it "returns :manual" do
-            expect(service.start_mode).to eq(:manual)
-          end
-        end
-
-        context "and there is no socket" do
-          let(:socket) { nil }
-
-          it "returns :manual" do
-            expect(service.start_mode).to eq(:manual)
-          end
-        end
-      end
-    end
-
-    describe "#start_mode=" do
-      subject(:service) { SystemdService.find("cups") }
-      let(:socket) { double("socket", disable: true) }
-
-      before do
-        stub_services(service: "cups")
-        allow(service).to receive(:socket).and_return(socket)
-        allow(service).to receive(:disable)
-      end
-
-      context "when :on_boot mode is given" do
-        it "enables the service to start on boot" do
-          expect(service).to receive(:enable)
-          expect(socket).to_not receive(:enable)
-          service.start_mode = :on_boot
-        end
-      end
-
-      context "when :on_demand mode is given" do
-        it "enables the socket" do
-          expect(service).to_not receive(:enable)
-          expect(socket).to receive(:enable)
-          service.start_mode = :on_demand
-        end
-      end
-
-      context "when :manual mode is given" do
-        it "disables the service and the socket" do
-          expect(service).to receive(:disable)
-          expect(socket).to receive(:disable)
-          service.start_mode = :manual
-        end
-      end
-
-      context "when an invalid value is given" do
-        it "raises an error" do
-          expect { service.start_mode = :other }.to raise_error(ArgumentError)
-        end
-      end
-    end
-
-    describe "#start_modes" do
-      subject(:service) { SystemdService.find("cups") }
-
-      before do
-        stub_services(service: "cups")
-        allow(service).to receive(:socket).and_return(socket)
-      end
-
-      context "when an associated socket exists" do
-        let(:socket) { double("socket", disable: true) }
-
-        it "returns :on_boot, :on_demand and :manual" do
-          expect(service.start_modes).to eq([:on_boot, :on_demand, :manual])
-        end
-      end
-
-      context "when no associated socket exists" do
-        let(:socket) { nil }
-
-        it "returns :on_boot and :manual" do
-          expect(service.start_modes).to eq([:on_boot, :manual])
-        end
-      end
-    end
-
-    describe "#socket?" do
-      subject(:service) { SystemdService.find("cups") }
-
-      before do
-        allow(service).to receive(:socket).and_return(socket)
-      end
-
-      context "when there is an associated socket" do
-        let(:socket) { double("socket") }
-
-        it "returns true" do
-          expect(service.socket?).to eq(true)
-        end
-      end
-
-      context "when there is no associated socket" do
-        let(:socket) { nil }
-
-        it "returns false" do
-          expect(service.socket?).to eq(false)
-        end
-      end
-    end
-
     context "Start a service on the installation system" do
       it "starts a service with a specialized inst-sys helper if available" do
         allow(File).to receive(:exist?).with("/bin/service_start").and_return(true)
